Move tests to unit directory
* salesforcedx-visualforce-markup-language-server package
