+++ packages/salesforcedx-visualforce-markup-language-server/test/unit/parser.test.ts
@@ -5,7 +5,7 @@
 'use strict';
 
 import * as assert from 'assert';
-import { HTMLDocument, Node, parse } from '../src/parser/htmlParser';
+import { HTMLDocument, Node, parse } from '../../src/parser/htmlParser';
 
 describe('HTML Parser', () => {
   function toJSON(node: Node) {
