+++ packages/salesforcedx-visualforce-markup-language-server/test/unit/symbols.test.ts
@@ -12,7 +12,7 @@ import {
   SymbolKind,
   TextDocument
 } from 'vscode-languageserver-types';
-import * as htmlLanguageService from '../src/htmlLanguageService';
+import * as htmlLanguageService from '../../src/htmlLanguageService';
 
 describe('HTML Symbols', () => {
   const TEST_URI = 'test://test/test.html';
