+++ packages/salesforcedx-visualforce-markup-language-server/test/unit/textEditSupport.ts
